requirements.txt
requests>=2.28.0
argparse>=1.4.0
pyperclip>=1.8.2
python-whois>=0.8.0
dnspython>=2.2.1
pandas>=1.5.0
Jinja2>=3.1.2
reportlab>=3.6.0
PyYAML>=6.0